02_build/covered-ai-v2/requirements.txt
# Covered AI v2 - Python Dependencies

# Core
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
prisma>=0.11.0

# Async HTTP
httpx>=0.26.0
aiohttp>=3.9.0

# Twilio
twilio>=8.10.0

# Email
resend>=0.7.0

# PDF Generation
reportlab>=4.0.0

# Date/Time utilities
python-dateutil>=2.8.0

# Payments
stripe>=8.0.0

# Production
gunicorn>=21.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dev
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
